Rename abstract controllers
