+++ app/Http/Controllers/BuildController.php
@@ -1,38 +1,12 @@
 <?php
 namespace App\Http\Controllers;
 
-use App\Services\TestingDay;
 use CDash\Database;
-use CDash\Model\Build;
-use App\Models\Site;
-use CDash\Model\Project;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\View\View;
 
-class BuildController extends ProjectController
+class BuildController extends AbstractBuildController
 {
-    protected Build $build;
-
-    // Fetch data used by all build-specific pages in CDash.
-    protected function setBuild(Build $build): void
-    {
-        if (!$build->Exists()) {
-            abort(404, 'Build does not exist. Maybe it has been deleted.');
-        }
-
-        $this->setProject($build->GetProject());
-        $this->build = $build;
-        $this->date = TestingDay::get($this->project, $this->build->StartTime);
-    }
-
-    protected function setBuildById(int $buildid): void
-    {
-        $build = new Build();
-        $build->Id = $buildid;
-        $build->FillFromId($buildid);
-        $this->setBuild($build);
-    }
-
     // Render the build configure page.
     public function configure($build_id = null)
     {
